HUGE CLEANUP
Mark unused variables as such
Explicitly invoke the parent;s constructor in copy constructors
Replace C includes with equivalent C++
Format
Remove non existent include folders from eclipse project files

+++ nettrader/ntbpsrc/ntbp_transition_rate_matrix_obj.cpp
@@ -10,7 +10,7 @@
 NTBP_transition_rate_matrix_o::NTBP_transition_rate_matrix_o(
 		NTsize new_num_states, NTreal new_min, NTreal new_max, NTreal new_step) :
 		min(new_min), max(new_max), step(new_step), num_states(new_num_states) {
-	long long length = (floor((max - min) / step + 0.5) + 1)	* num_states * num_states;
+	int length = (floor((max - min) / step + 0.5) + 1)	* num_states * num_states;
 	_probMatrices = new NTreal[length];
 	for (int i = 0; i< length; i++){
 		_probMatrices[i]=0;
